requirements.txt
coincurve
pycryptodome